+++ Assembly/Whiskey/RequiresToolAttribute.cs
@@ -7,7 +7,18 @@ namespace Whiskey
         public RequiresToolAttribute(string toolName)
         {
             Name = toolName;
+            var nameStartsAt = toolName.IndexOf("::");
+            if( nameStartsAt >= 0 )
+            {
+                ProviderName = toolName.Substring(0, nameStartsAt);
+                Name = toolName.Substring(nameStartsAt + 2);
+            }
             VersionParameterName = "Version";
+        }
+
+        public RequiresToolAttribute(string toolName, string providerName) : this(toolName)
+        {
+            ProviderName = providerName;
         }
 
         public bool AddToPath { get; set; }
@@ -15,6 +26,8 @@ namespace Whiskey
         public string Name { get; private set; }
 
         public string PathParameterName { get; set; }
+
+        public string ProviderName { get; set; }
 
         public string Version { get; set; }