+++ Test/NuGetPush.Tests.ps1
@@ -4,6 +4,10 @@ Set-StrictMode -Version 'Latest'
 
 & (Join-Path -Path $PSScriptRoot -ChildPath 'Initialize-WhiskeyTest.ps1' -Resolve)
 
+# These tests intermittently fail. Turn on verbose and debug to maybe help locate the problem.
+$DebugPreference = [Management.Automation.ActionPreference]::Continue
+$VerbosePreference = [Management.Automation.ActionPreference]::Continue
+
 $testRoot = $null
 $context = $null
 $nugetUri = $null
@@ -200,8 +204,8 @@ function WhenRunningNuGetPackTask
             $taskParameter['Path'] = 'I\do\not\exist.csproj'
         }
 
-        $Global:error.Clear()
-        Invoke-WhiskeyTask -TaskContext $context -Parameter $taskParameter -Name 'NuGetPush' | Out-Null 
+        $Global:Error.Clear()
+        Invoke-WhiskeyTask -TaskContext $context -Parameter $taskParameter -Name 'NuGetPush'
 
     }
     catch