+++ Whiskey/Tasks/NodeLicenseChecker.ps1
@@ -3,8 +3,8 @@ function Invoke-WhiskeyNodeLicenseChecker
 {
     [CmdletBinding()]
     [Whiskey.Task('NodeLicenseChecker')]
-    [Whiskey.RequiresTool('Node',PathParameterName='NodePath',VersionParameterName='NodeVersion')]
-    [Whiskey.RequiresTool('NodeModule::license-checker',PathParameterName='LicenseCheckerPath',VersionParameterName='Version')]
+    [Whiskey.RequiresTool('Node', PathParameterName='NodePath', VersionParameterName='NodeVersion')]
+    [Whiskey.RequiresNodeModule('license-checker', PathParameterName='LicenseCheckerPath')]
     param(
         [Parameter(Mandatory)]
         [Whiskey.Context]$TaskContext,
@@ -17,6 +17,7 @@ function Invoke-WhiskeyNodeLicenseChecker
 
     Set-StrictMode -Version 'Latest'
     Use-CallerPreference -Cmdlet $PSCmdlet -SessionState $ExecutionContext.SessionState
+
     $licenseCheckerPath = Assert-WhiskeyNodeModulePath -Path $TaskParameter['LicenseCheckerPath'] -CommandPath 'bin\license-checker' -ErrorAction Stop
 
     $nodePath = Assert-WhiskeyNodePath -Path $TaskParameter['NodePath'] -ErrorAction Stop